Add HTTP-triggered Azure Function with anonymous access
Defined HTTP trigger and output bindings in `function.json`
to enable an Azure Function that responds to GET requests.
Implemented the `Run` method in `run.csx` to log requests
and return a "Hello from Azure Function!" message. This
function allows unauthenticated access and serves as a
lightweight serverless HTTP endpoint.